requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Pillow==10.2.0
opencv-python>=4.8.0
requests==2.31.0
notion-client==2.2.1
boto3==1.34.69
openai>=1.0.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
pytesseract>=0.3.10
numpy>=1.24.0
tiktoken>=0.5.0